+++ packages/api/tests/serial/endpoints/granules.js
@@ -1,9 +1,11 @@
 'use strict';
 
 const fs = require('fs');
+const request = require('supertest');
 const path = require('path');
 const sinon = require('sinon');
 const test = require('ava');
+const { sfn } = require('@cumulus/common/aws');
 const aws = require('@cumulus/common/aws');
 const { CMR } = require('@cumulus/cmrjs');
 const {
@@ -18,7 +20,6 @@ const { xmlParseOptions } = require('@cumulus/cmrjs/utils');
 const assertions = require('../../../lib/assertions');
 const models = require('../../../models');
 const bootstrap = require('../../../lambdas/bootstrap');
-const handleRequest = require('../../../endpoints/granules');
 const indexer = require('../../../es/indexer');
 const {
   fakeAccessTokenFactory,
@@ -30,6 +31,17 @@ const {
   createJwtToken
 } = require('../../../lib/token');
 const { Search } = require('../../../es/search');
+
+process.env.AccessTokensTable = randomString();
+process.env.CollectionsTable = randomString();
+process.env.GranulesTable = randomString();
+process.env.UsersTable = randomString();
+process.env.stackName = randomString();
+process.env.system_bucket = randomString();
+process.env.TOKEN_SECRET = randomString();
+
+// import the express app after setting the env variables
+const { app } = require('../../../app');
 
 const createBucket = (Bucket) => aws.s3().createBucket({ Bucket }).promise();
 
@@ -94,13 +106,6 @@ let userModel;
 
 test.before(async () => {
   esIndex = randomString();
-  process.env.AccessTokensTable = randomId('accesstable');
-  process.env.CollectionsTable = randomId('collectiontable');
-  process.env.GranulesTable = randomId('granulestable');
-  process.env.UsersTable = randomId('userstable');
-  process.env.stackName = randomId('stackname');
-  process.env.internal = randomId('internal');
-  process.env.TOKEN_SECRET = randomId('token');
 
   // create esClient
   esClient = await Search.es('fakehost');
@@ -109,7 +114,7 @@ test.before(async () => {
   await bootstrap.bootstrapElasticSearch('fakehost', esIndex);
 
   // create a fake bucket
-  await createBucket(process.env.internal);
+  await createBucket(process.env.system_bucket);
 
   // create fake Collections table
   collectionModel = new models.Collection();
@@ -130,10 +135,6 @@ test.before(async () => {
 });
 
 test.beforeEach(async (t) => {
-  t.context.authHeaders = {
-    Authorization: `Bearer ${accessToken}`
-  };
-
   t.context.testCollection = fakeCollectionFactory({
     name: 'fakeCollection',
     dataType: 'fakeCollection',
@@ -159,18 +160,17 @@ test.after.always(async () => {
   await accessTokenModel.deleteTable();
   await userModel.deleteTable();
   await esClient.indices.delete({ index: esIndex });
-  await aws.recursivelyDeleteS3Bucket(process.env.internal);
+  await aws.recursivelyDeleteS3Bucket(process.env.system_bucket);
 });
 
 test.serial('default returns list of granules', async (t) => {
-  const event = {
-    httpMethod: 'GET',
-    headers: t.context.authHeaders
-  };
+  const response = await request(app)
+    .get('/granules')
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .expect(200);
 
-  const response = await handleRequest(event);
-
-  const { meta, results } = JSON.parse(response.body);
+  const { meta, results } = response.body;
   t.is(results.length, 2);
   t.is(meta.stack, process.env.stackName);
   t.is(meta.table, 'granule');
@@ -182,67 +182,47 @@ test.serial('default returns list of granules', async (t) => {
 });
 
 test.serial('CUMULUS-911 GET without pathParameters and without an Authorization header returns an Authorization Missing response', async (t) => {
-  const request = {
-    httpMethod: 'GET',
-    headers: {}
-  };
-
-  const response = await handleRequest(request);
+  const response = await request(app)
+    .get('/granules')
+    .set('Accept', 'application/json')
+    .expect(401);
 
   assertions.isAuthorizationMissingResponse(t, response);
 });
 
 test.serial('CUMULUS-911 GET with pathParameters.granuleName set and without an Authorization header returns an Authorization Missing response', async (t) => {
-  const request = {
-    httpMethod: 'GET',
-    headers: {},
-    pathParameters: {
-      granuleName: 'asdf'
-    }
-  };
-
-  const response = await handleRequest(request);
+  const response = await request(app)
+    .get('/granules/asdf')
+    .set('Accept', 'application/json')
+    .expect(401);
 
   assertions.isAuthorizationMissingResponse(t, response);
 });
 
 test.serial('CUMULUS-911 PUT with pathParameters.granuleName set and without an Authorization header returns an Authorization Missing response', async (t) => {
-  const request = {
-    httpMethod: 'PUT',
-    headers: {},
-    pathParameters: {
-      granuleName: 'asdf'
-    }
-  };
-
-  const response = await handleRequest(request);
+  const response = await request(app)
+    .put('/granules/asdf')
+    .set('Accept', 'application/json')
+    .expect(401);
 
   assertions.isAuthorizationMissingResponse(t, response);
 });
 
 test.serial('CUMULUS-911 DELETE with pathParameters.granuleName set and without an Authorization header returns an Authorization Missing response', async (t) => {
-  const request = {
-    httpMethod: 'DELETE',
-    headers: {},
-    pathParameters: {
-      granuleName: 'asdf'
-    }
-  };
-
-  const response = await handleRequest(request);
+  const response = await request(app)
+    .delete('/granules/asdf')
+    .set('Accept', 'application/json')
+    .expect(401);
 
   assertions.isAuthorizationMissingResponse(t, response);
 });
 
 test.serial('CUMULUS-912 GET without pathParameters and with an invalid access token returns an unauthorized response', async (t) => {
-  const request = {
-    httpMethod: 'GET',
-    headers: {
-      Authorization: 'Bearer ThisIsAnInvalidAccessToken'
-    }
-  };
-
-  const response = await handleRequest(request);
+  const response = await request(app)
+    .get('/granules/asdf')
+    .set('Accept', 'application/json')
+    .set('Authorization', 'Bearer ThisIsAnInvalidAuthorizationToken')
+    .expect(403);
 
   assertions.isInvalidAccessTokenResponse(t, response);
 });
@@ -252,30 +232,21 @@ test.serial('CUMULUS-912 GET without pathParameters and with an unauthorized use
   await accessTokenModel.create(accessTokenRecord);
   const jwtToken = createJwtToken(accessTokenRecord);
 
-  const request = {
-    httpMethod: 'GET',
-    headers: {
-      Authorization: `Bearer ${jwtToken}`
-    }
-  };
-
-  const response = await handleRequest(request);
+  const response = await request(app)
+    .get('/granules')
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${jwtToken}`)
+    .expect(401);
 
   assertions.isUnauthorizedUserResponse(t, response);
 });
 
 test.serial('CUMULUS-912 GET with pathParameters.granuleName set and with an invalid access token returns an unauthorized response', async (t) => {
-  const request = {
-    httpMethod: 'GET',
-    headers: {
-      Authorization: 'Bearer ThisIsAnInvalidAuthorizationToken'
-    },
-    pathParameters: {
-      granuleName: 'asdf'
-    }
-  };
-
-  const response = await handleRequest(request);
+  const response = await request(app)
+    .get('/granules/asdf')
+    .set('Accept', 'application/json')
+    .set('Authorization', 'Bearer ThisIsAnInvalidAuthorizationToken')
+    .expect(403);
 
   assertions.isInvalidAccessTokenResponse(t, response);
 });
@@ -283,17 +254,11 @@ test.serial('CUMULUS-912 GET with pathParameters.granuleName set and with an inv
 test.todo('CUMULUS-912 GET with pathParameters.granuleName set and with an unauthorized user returns an unauthorized response');
 
 test.serial('CUMULUS-912 PUT with pathParameters.granuleName set and with an invalid access token returns an unauthorized response', async (t) => {
-  const request = {
-    httpMethod: 'PUT',
-    headers: {
-      Authorization: 'Bearer ThisIsAnInvalidAuthorizationToken'
-    },
-    pathParameters: {
-      granuleName: 'asdf'
-    }
-  };
-
-  const response = await handleRequest(request);
+  const response = await request(app)
+    .put('/granules/asdf')
+    .set('Accept', 'application/json')
+    .set('Authorization', 'Bearer ThisIsAnInvalidAuthorizationToken')
+    .expect(403);
 
   assertions.isInvalidAccessTokenResponse(t, response);
 });
@@ -305,82 +270,60 @@ test.serial('CUMULUS-912 DELETE with pathParameters.granuleName set and with an 
   await accessTokenModel.create(accessTokenRecord);
   const jwtToken = createJwtToken(accessTokenRecord);
 
-  const request = {
-    httpMethod: 'DELETE',
-    headers: {
-      Authorization: `Bearer ${jwtToken}`
-    },
-    pathParameters: {
-      granuleName: 'asdf'
-    }
-  };
-
-  const response = await handleRequest(request);
+  const response = await request(app)
+    .delete('/granules/adsf')
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${jwtToken}`)
+    .expect(401);
 
   assertions.isUnauthorizedUserResponse(t, response);
 });
 
 test.serial('GET returns an existing granule', async (t) => {
-  const event = {
-    httpMethod: 'GET',
-    pathParameters: {
-      granuleName: t.context.fakeGranules[0].granuleId
-    },
-    headers: t.context.authHeaders
-  };
+  const response = await request(app)
+    .get(`/granules/${t.context.fakeGranules[0].granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .expect(200);
 
-  const response = await handleRequest(event);
-
-  const { granuleId } = JSON.parse(response.body);
+  const { granuleId } = response.body;
   t.is(granuleId, t.context.fakeGranules[0].granuleId);
 });
 
 test.serial('GET returns a 404 response if the granule is not found', async (t) => {
-  const event = {
-    httpMethod: 'GET',
-    pathParameters: {
-      granuleName: 'unknownGranule'
-    },
-    headers: t.context.authHeaders
-  };
+  const response = await request(app)
+    .get('/granules/unknownGranule')
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .expect(404);
 
-  const response = await handleRequest(event);
-
-  t.is(response.statusCode, 404);
-  const { message } = JSON.parse(response.body);
+  t.is(response.status, 404);
+  const { message } = response.body;
   t.is(message, 'Granule not found');
 });
 
 test.serial('PUT fails if action is not supported', async (t) => {
-  const event = {
-    httpMethod: 'PUT',
-    pathParameters: {
-      granuleName: t.context.fakeGranules[0].granuleId
-    },
-    headers: t.context.authHeaders,
-    body: JSON.stringify({ action: 'SomeUnsuportedAction' })
-  };
+  const response = await request(app)
+    .put(`/granules/${t.context.fakeGranules[0].granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .send({ action: 'someUnsupportedAction' })
+    .expect(400);
 
-  const response = await handleRequest(event);
-
-  t.is(response.statusCode, 400);
-  const { message } = JSON.parse(response.body);
+  t.is(response.status, 400);
+  const { message } = response.body;
   t.true(message.includes('Action is not supported'));
 });
 
 test.serial('PUT fails if action is not provided', async (t) => {
-  const event = {
-    httpMethod: 'PUT',
-    pathParameters: {
-      granuleName: t.context.fakeGranules[0].granuleId
-    },
-    headers: t.context.authHeaders
-  };
+  const response = await request(app)
+    .put(`/granules/${t.context.fakeGranules[0].granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .expect(400);
 
-  const response = await handleRequest(event);
-
-  t.is(response.statusCode, 400);
-  const { message } = JSON.parse(response.body);
+  t.is(response.status, 400);
+  const { message } = response.body;
   t.is(message, 'Action is missing');
 });
 
@@ -395,42 +338,29 @@ test.serial('reingest a granule', async (t) => {
     })
   };
 
-  const event = {
-    httpMethod: 'PUT',
-    pathParameters: {
-      granuleName: t.context.fakeGranules[0].granuleId
-    },
-    headers: t.context.authHeaders,
-    body: JSON.stringify({ action: 'reingest' })
-  };
-
   // fake workflow
-  process.env.bucket = process.env.internal;
   const message = JSON.parse(fakeDescribeExecutionResult.input);
   const key = `${process.env.stackName}/workflows/${message.meta.workflow_name}.json`;
-  await putObject({ Bucket: process.env.bucket, Key: key, Body: 'test data' });
+  await putObject({ Bucket: process.env.system_bucket, Key: key, Body: 'test data' });
+  const stub = sinon.stub(sfn(), 'describeExecution').returns({
+    promise: () => Promise.resolve(fakeDescribeExecutionResult)
+  });
 
-  const sfn = aws.sfn();
+  const response = await request(app)
+    .put(`/granules/${t.context.fakeGranules[0].granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .send({ action: 'reingest' })
+    .expect(200);
 
-  let response;
-  try {
-    sfn.describeExecution = () => ({
-      promise: () => Promise.resolve(fakeDescribeExecutionResult)
-    });
-
-    response = await handleRequest(event);
-  }
-  finally {
-    delete sfn.describeExecution;
-  }
-
-  const body = JSON.parse(response.body);
+  const body = response.body;
   t.is(body.status, 'SUCCESS');
   t.is(body.action, 'reingest');
   t.true(body.warning.includes('overwritten'));
 
   const updatedGranule = await granuleModel.get({ granuleId: t.context.fakeGranules[0].granuleId });
   t.is(updatedGranule.status, 'running');
+  stub.restore();
 });
 
 // This needs to be serial because it is stubbing aws.sfn's responses
@@ -446,24 +376,10 @@ test.serial('apply an in-place workflow to an existing granule', async (t) => {
     }
   };
 
-  const event = {
-    httpMethod: 'PUT',
-    pathParameters: {
-      granuleName: t.context.fakeGranules[0].granuleId
-    },
-    headers: t.context.authHeaders,
-    body: JSON.stringify({
-      action: 'applyWorkflow',
-      workflow: 'inPlaceWorkflow',
-      messageSource: 'output'
-    })
-  };
-
   //fake in-place workflow
-  process.env.bucket = process.env.internal;
   const message = JSON.parse(fakeSFResponse.execution.input);
   const key = `${process.env.stackName}/workflows/${message.meta.workflow_name}.json`;
-  await putObject({ Bucket: process.env.bucket, Key: key, Body: 'fake in-place workflow' });
+  await putObject({ Bucket: process.env.system_bucket, Key: key, Body: 'fake in-place workflow' });
 
   const fakeDescribeExecutionResult = {
     output: JSON.stringify({
@@ -474,38 +390,31 @@ test.serial('apply an in-place workflow to an existing granule', async (t) => {
     })
   };
 
-  const sfn = aws.sfn();
+  const stub = sinon.stub(sfn(), 'describeExecution').returns({
+    promise: () => Promise.resolve(fakeDescribeExecutionResult)
+  });
 
-  let response;
-  try {
-    sfn.describeExecution = () => ({
-      promise: () => Promise.resolve(fakeDescribeExecutionResult)
-    });
+  const response = await request(app)
+    .put(`/granules/${t.context.fakeGranules[0].granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .send({
+      action: 'applyWorkflow',
+      workflow: 'inPlaceWorkflow',
+      messageSource: 'output'
+    })
+    .expect(200);
 
-    response = await handleRequest(event);
-  }
-  finally {
-    delete sfn.describeExecution;
-  }
-
-  const body = JSON.parse(response.body);
+  const body = response.body;
   t.is(body.status, 'SUCCESS');
   t.is(body.action, 'applyWorkflow inPlaceWorkflow');
 
   const updatedGranule = await granuleModel.get({ granuleId: t.context.fakeGranules[0].granuleId });
   t.is(updatedGranule.status, 'running');
+  stub.restore();
 });
 
 test.serial('remove a granule from CMR', async (t) => {
-  const event = {
-    httpMethod: 'PUT',
-    pathParameters: {
-      granuleName: t.context.fakeGranules[0].granuleId
-    },
-    headers: t.context.authHeaders,
-    body: JSON.stringify({ action: 'removeFromCmr' })
-  };
-
   sinon.stub(
     DefaultProvider,
     'decrypt'
@@ -516,33 +425,35 @@ test.serial('remove a granule from CMR', async (t) => {
     'deleteGranule'
   ).callsFake(() => Promise.resolve());
 
-  const response = await handleRequest(event);
+  const response = await request(app)
+    .put(`/granules/${t.context.fakeGranules[0].granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .send({ action: 'removeFromCmr' })
+    .expect(200);
 
-  const body = JSON.parse(response.body);
+
+  const body = response.body;
   t.is(body.status, 'SUCCESS');
   t.is(body.action, 'removeFromCmr');
 
   const updatedGranule = await granuleModel.get({ granuleId: t.context.fakeGranules[0].granuleId });
   t.is(updatedGranule.published, false);
-  t.is(updatedGranule.cmrLink, null);
+  t.is(updatedGranule.cmrLink, undefined);
 
   CMR.prototype.deleteGranule.restore();
   DefaultProvider.decrypt.restore();
 });
 
 test.serial('DELETE deleting an existing granule that is published will fail', async (t) => {
-  const event = {
-    httpMethod: 'DELETE',
-    pathParameters: {
-      granuleName: t.context.fakeGranules[0].granuleId
-    },
-    headers: t.context.authHeaders
-  };
+  const response = await request(app)
+    .delete(`/granules/${t.context.fakeGranules[0].granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .expect(400);
 
-  const response = await handleRequest(event);
-
-  t.is(response.statusCode, 400);
-  const { message } = JSON.parse(response.body);
+  t.is(response.status, 400);
+  const { message } = response.body;
   t.is(
     message,
     'You cannot delete a granule that is published to CMR. Remove it from CMR first'
@@ -598,18 +509,14 @@ test.serial('DELETE deleting an existing unpublished granule', async (t) => {
   // create a new unpublished granule
   await granuleModel.create(newGranule);
 
-  const event = {
-    httpMethod: 'DELETE',
-    pathParameters: {
-      granuleName: newGranule.granuleId
-    },
-    headers: t.context.authHeaders
-  };
+  const response = await request(app)
+    .delete(`/granules/${newGranule.granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .expect(200);
 
-  const response = await handleRequest(event);
-
-  t.is(response.statusCode, 200);
-  const { detail } = JSON.parse(response.body);
+  t.is(response.status, 200);
+  const { detail } = response.body;
   t.is(detail, 'Record deleted');
 
   // verify the files are deleted
@@ -628,7 +535,7 @@ test.serial('DELETE deleting an existing unpublished granule', async (t) => {
 });
 
 test.serial('move a granule with no .cmr.xml file', async (t) => {
-  const bucket = process.env.internal;
+  const bucket = process.env.system_bucket;
   const secondBucket = randomId('second');
   const thirdBucket = randomId('third');
 
@@ -683,21 +590,17 @@ test.serial('move a granule with no .cmr.xml file', async (t) => {
         }
       ];
 
-      const event = {
-        httpMethod: 'PUT',
-        pathParameters: {
-          granuleName: newGranule.granuleId
-        },
-        headers: t.context.authHeaders,
-        body: JSON.stringify({
+      const response = await request(app)
+        .put(`/granules/${newGranule.granuleId}`)
+        .set('Accept', 'application/json')
+        .set('Authorization', `Bearer ${accessToken}`)
+        .send({
           action: 'move',
           destinations
         })
-      };
+        .expect(200);
 
-      const response = await handleRequest(event);
-
-      const body = JSON.parse(response.body);
+      const body = response.body;
       t.is(body.status, 'SUCCESS');
       t.is(body.action, 'move');
 
@@ -774,26 +677,22 @@ test.serial('move a file and update ECHO10 xml metadata', async (t) => {
     }
   ];
 
-  const event = {
-    httpMethod: 'PUT',
-    pathParameters: {
-      granuleName: newGranule.granuleId
-    },
-    headers: t.context.authHeaders,
-    body: JSON.stringify({
-      action: 'move',
-      destinations
-    })
-  };
-
   sinon.stub(
     CMR.prototype,
     'ingestGranule'
   ).returns({ result: { 'concept-id': 'id204842' } });
 
-  const response = await handleRequest(event);
+  const response = await request(app)
+    .put(`/granules/${newGranule.granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .send({
+      action: 'move',
+      destinations
+    })
+    .expect(200);
 
-  const body = JSON.parse(response.body);
+  const body = response.body;
 
   t.is(body.status, 'SUCCESS');
   t.is(body.action, 'move');
@@ -928,19 +827,16 @@ test('PUT with action move returns failure if one granule file exists', async (t
     }]
   };
 
-  const request = {
-    httpMethod: 'PUT',
-    pathParameters: {
-      granuleName: granule.granuleId
-    },
-    body: JSON.stringify(body),
-    headers: t.context.authHeaders
-  };
+  const response = await request(app)
+    .put(`/granules/${granule.granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .send(body)
+    .expect(409);
 
-  const response = await handleRequest(request);
 
-  const responseBody = JSON.parse(response.body);
-  t.is(response.statusCode, 409);
+  const responseBody = response.body;
+  t.is(response.status, 409);
   t.is(responseBody.message,
     'Cannot move granule because the following files would be overwritten at the destination location: file1. Delete the existing files or reingest the source files.');
 
@@ -969,18 +865,14 @@ test('PUT with action move returns failure if more than one granule file exists'
     }]
   };
 
-  const request = {
-    httpMethod: 'PUT',
-    pathParameters: {
-      granuleName: granule.granuleId
-    },
-    body: JSON.stringify(body),
-    headers: t.context.authHeaders
-  };
+  const response = await request(app)
+    .put(`/granules/${granule.granuleId}`)
+    .set('Accept', 'application/json')
+    .set('Authorization', `Bearer ${accessToken}`)
+    .send(body)
+    .expect(409);
 
-  const response = await handleRequest(request);
-
-  const responseBody = JSON.parse(response.body);
+  const responseBody = response.body;
   t.is(response.statusCode, 409);
   t.is(responseBody.message,
     'Cannot move granule because the following files would be overwritten at the destination location: file1, file2, file3. Delete the existing files or reingest the source files.');