+++ packages/common/test-utils.js
@@ -49,6 +49,7 @@ exports.randomStringFromRegex = (regex) => new RandExp(regex).gen();
 
 // From https://github.com/localstack/localstack/blob/master/README.md
 const localStackPorts = {
+  stepfunctions: 10000, // add a fake port to support test overrides
   apigateway: 4567,
   cloudformation: 4581,
   cloudwatch: 4582,